Upgrade CI to latest Ubuntu LTS
