+++ package.json
@@ -116,7 +116,6 @@
         "karma-sinon-chai": "^2.0.2",
         "mocha": "^9.0.2",
         "nock": "^13.1.1",
-        "object.fromentries": "^2.0.4",
         "rimraf": "^3.0.2",
         "sinon": "^11.1.1",
         "sinon-chai": "^3.7.0",
