--- setup/travis/redis-enketo-cache.conf
@@ -1,15 +1,0 @@
-# redis-server - Enketo Main Instance
-#
-# Redis is a key value in memory persistent datastore
-
-start on (local-filesystems and runlevel [2345])
-stop on runlevel [016]
-respawn
-expect fork
-limit nofile 20000 65000
-pre-start script
-mkdir -p /var/run/redis
-chown redis:redis /var/run/redis
-end script
-
-exec start-stop-daemon --start --chuid redis:redis --pidfile /var/run/redis-enketo-cache/redis-server.pid --umask 007 --exec /usr/bin/redis-server -- /etc/redis/redis-enketo-cache.conf
