+++ test/client/connection.spec.js
@@ -9,17 +9,6 @@ import settings from '../../public/js/src/module/settings';
 import store from '../../public/js/src/module/store';
 
 /**
- * @see {@link https://github.com/enketo/enketo-express/pull/269#issuecomment-861887583}
- * TODO [2021-07-22]: remove these polyfills when CI is able to use a newer version of Chrome
- */
-import 'abortcontroller-polyfill/dist/polyfill-patch-fetch';
-import fromEntries from 'object.fromentries';
-if ( Object.fromEntries == null ) {
-    fromEntries.shim();
-}
-
-
-/**
  * @typedef {import('../../app/models/record-model').EnketoRecord} EnketoRecord
  */
